+++ src/screens/CartScreen/CartScreen.tsx
@@ -62,12 +62,12 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
   const { userData } = userDataStore;
 
   // TODO: don't use any type
-  // TODO: resolve item unit
   const formatAccrodion = (data: any[]): AccrodionPriceModel[] => {
     let arrayOutput: any[] = [];
     data.map((item: any) => {
+      let unit = item.sale_unit ? item.sale_unit : item.unit
       arrayOutput.push({
-        item: `${item.name} (${item.price}฿ x ${item.quantity} ลัง)`,
+        item: `${item.name} (${item.price}฿ x ${item.quantity} ${unit})`,
         price: item.price,
         quantity: item.quantity,
       });