+++ src/screens/OrderSuccessScreen/OrderSuccessScreen.tsx
@@ -38,8 +38,9 @@ const OrderSuccessScreen: React.FC<OrderSuccessScreenRouteProp> = ({
   const formatAccrodion = (data: any[]): AccrodionPriceModel[] => {
     let arrayOutput: any[] = [];
     data.map((item: any) => {
+      let unit = item.sale_unit ? item.sale_unit : item.unit
       arrayOutput.push({
-        item: `${item.name} (${item.price}฿ x ${item.quantity} ลัง)`,
+        item: `${item.name} (${item.price}฿ x ${item.quantity} $unit)`,
         price: item.price,
         quantity: item.quantity,
       });