+++ src/screens/OrderSuccessScreenDetail/OrderSuccessScreenDetail.tsx
@@ -70,12 +70,12 @@ const OrderSuccessScreenDetail: React.FC<OrderSuccessScreenDetailRouteProp> = ({
   };
 
   // TODO: don't use any type
-  // TODO: resolve item unit
   const formatAccrodion = (data: any[]): AccrodionPriceModel[] => {
     let arrayOutput: any[] = [];
     data.map((item: any) => {
+      let unit = item.sale_unit ? item.sale_unit : item.unit
       arrayOutput.push({
-        item: `${item.name} (${item.price}฿ x ${item.quantity} ลัง)`,
+        item: `${item.name} (${item.price}฿ x ${item.quantity} ${unit})`,
         price: item.price,
         quantity: item.quantity,
       });